added last_bias_initializer arg  and sigmoid vs softmax option at last layer
This optional parameter can be used to start the model training off with a custom bias term in the last layer in order to reflect the imbalance between the two classes, and expedite training and perhaps improve training results. 
Also, I added an option to switch between a softmax and sigmoid in the last activation function, which decides whether one is doing a multiple class segmentation (of two or more distinct classes, which entails 2 or more mask channels) or a binary segmentation (of one class, which effectively means the class pixels are of the class and the other pixels are not-of-the-class, this sounds like to classes, but it's not because each pixel is either or, and can't be both, also the segmentation mask channel can only have one channel).